requirements.txt
langchain
streamlit
faiss-cpu
python-dotenv
tiktoken
gemini
fireworks
sentence_transformers